Prevent Multiple Autocomplete Library Imports
This is only a problem when integrating with other Google Maps libraries
as the current logic only ensures that additional instances of this component
don't import the same library.

While this patch attempted to take into consideration scripts originating elsewhere,
this implementation does not fully address the problem because other libraries can
still insert duplicate location autocomplete libraries dynamically (therefore triggering
the same warning).  It is also the case if Google Maps script is inserted later in the DOM, as this component will have already inserted the script tag to the header.

In other words, this patch would only work if the Google Maps library is included before this component is mounted.

Recommendation:
For the time being, it is not advised to use `location-autocomplete` with another Google Maps integration as it could result in the following warning:
```
You have included the Google Maps API multiple times on this page. This may cause unexpected errors.
```

Issue: #10